==> notebooks/0_exploritory_notebook.ipynb <==
from owid import catalog
from owid.catalog import charts

# opening TheWorldBanks's catalogue (through OurWordlInData API) to see what kind of datasets they have related to "plastic"
plastic_catalogs = catalog.find('plastic')
plastic_catalogs.table.unique()

Let's explore the datasets under the names "plastic_aquatic_leakage", "plastic_emissions", "plastic_fate_regions", "plastic_use_application", "plastic_use_polymer", "plastic_waste" & "plastic_waste_2019" as those are the ones that might conatain interesting insights into our research.

Let's begin with:
### Plastic Waste dataset

# loading up and taking a first look into some of the datasets 
dataset_name = "plastic_waste"
dataset_path = plastic_catalogs[plastic_catalogs['table'] == dataset_name].path.iloc[0]

# loading method from the docs
rc = catalog.RemoteCatalog() 
plastic_waste_df = rc[dataset_path]

plastic_catalogs[plastic_catalogs['table'] == dataset_name]

# metadata for the catalog 
plastic_waste_df.metadata

plastic_waste_df

print(f'Columns inside a dataset: {plastic_waste_df.columns}')
print(f'Years in the dataset: {plastic_waste_df.index.get_level_values(0).unique()}')
print(f'Countries count: {plastic_waste_df.index.get_level_values(1).nunique()}')
print(f'Countries: {plastic_waste_df.index.get_level_values(1).unique()}')

From the metadata and https://ourworldindata.org/grapher/global-plastics-production it seems that we are interested in the first two columns primarily ('import_total_mot' and 'export_total_mot'). They show the ammount million metric tones (Mt) of plastic imported and exported by countries. It is of interest, but first let's try to find the data on how much plastic is produced. Lets try the following dataset next:

### Plastic Use Polymer dataset

# loading up and taking a first look into some of the datasets 
dataset_name = "plastic_use_polymer"
dataset_path = plastic_catalogs[plastic_catalogs['table'] == dataset_name].path.iloc[0]

# loading method from the docs
rc = catalog.RemoteCatalog() 
plastic_polymer_df = rc[dataset_path]

plastic_catalogs[plastic_catalogs['table'] == dataset_name]

# metadata for the catalog 
plastic_polymer_df.metadata

plastic_polymer_df

print(f'Columns inside a dataset: {plastic_polymer_df.columns}')
print(f'Years in the dataset: {plastic_polymer_df.index.get_level_values(1).unique()}')
print(f'Countries count: {plastic_polymer_df.index.get_level_values(0).nunique()}')
print(f'Countries: {plastic_polymer_df.index.get_level_values(0).unique()}')

import pandas as pd
world_production = pd.read_csv('../data/plastic-waste-generation-total.csv')

world_production
