==> notebooks/1_total_plastic_waste.ipynb <==
import pandas as pd

from owid import catalog
from owid.catalog import charts

country_keys = pd.read_csv('../data/plastic-waste-generation-total.csv')
country_keys.columns = ['country', 'code', 'year', 'total_plastic_waste']
country_keys

# opening TheWorldBanks's catalogue
plastic_catalogs = catalog.find('plastic')
dataset_name = "plastic_use_polymer"
dataset_path = plastic_catalogs[plastic_catalogs['table'] == dataset_name].path.iloc[0]

# loading method from the docs
rc = catalog.RemoteCatalog() 
plastic_polymer_df = rc[dataset_path]

plastic_polymer_df.loc['World', 2010]

print(country_keys.country.nunique())
print(country_keys.year.unique())
print(country_keys.columns)
print(country_keys['total_plastic_waste'].sum())


total_plastic_polymere_production = plastic_polymer_df.loc['World', 2010]['total']
total_plastic_waste_generation = country_keys['total_plastic_waste'].sum()
waste_to_production = total_plastic_waste_generation/total_plastic_polymere_production
print(f'Total % of plastic produced turned to waste: {waste_to_production*100}')

plastic_polymer_df['total_to_2010'] = plastic_polymer_df['total'] / total_plastic_polymere_production
plastic_polymer_df = plastic_polymer_df.reset_index()
plastic_polymer_df = plastic_polymer_df.drop(columns='country')

columns_to_drop = [column for column in plastic_polymer_df.columns if column not in (['year', 'total', 'total_to_2010'])]
plastic_polymer_df = plastic_polymer_df.drop(columns=columns_to_drop)
plastic_polymer_df.head()

# expanded_df = pd.DataFrame()
# for country in country_keys['country'].unique():
#     full_df = pd.merge(plastic_polymer_df, country_keys[country_keys['country'] == country], on='year', how ='left')
#     full_df['country'] = country
#     full_df['total'] = full_df['total'] * full_df['total_to_2010']
#     expanded_df = pd.concat([expanded_df, full_df], ignore_index=True)



# columns_to_drop = [column for column in expanded_df.columns if column not in (['year', 'total', 'total_to_2010', 'country'])]
# expanded_df = expanded_df.drop(columns=columns_to_drop)

# adjust country keys with the coefficient waste_to_production to get the number of plastic produced
# ASSUMPTION 1
country_keys['total_plastic_waste'] /= waste_to_production
country_keys.columns = ['country', 'code', 'year', 'total_plastic_production']

extended_df = pd.DataFrame() 
for country in country_keys['country'].unique():
    working_country = country_keys[country_keys['country'] == country]
    a = plastic_polymer_df.copy()
    a['country_produced_plastic'] = a['total_to_2010'] * working_country['total_plastic_production'].iloc[0]
    a['country'] = country
    a = a.drop(columns=['total', 'total_to_2010'])
    extended_df = pd.concat([extended_df, a])


extended_df['country_produced_plastic'].nunique()

extended_df.to_csv('../data/01_world_spatio_temporal_plastic_production_v0.csv')
extended_df

* What I did here, is I extracted dataset with total polymere production in 1990-2019 (worldwide), and extracted another dataset with 168 countries data with plastic waste generation in 2010.
* Then I calculated the total ammount of plastic waste generated (ASSUMPTION 1: the world consist of 168 countries and no more), and got a proportion of plastic to waste generation. To get the approximate number of plastic that each country produced in 2010. 
* To get the country production by year, I saw how much the prodiction in the first dataset was changing (in proportion to 2010) and applied these proportions to every country

TODO: 
* [] try to find the data for each country's GDP 
* [] adjust total production of polymers to the proprtion of this countries GDP's production (details are in the notes)
* [] correct the number of yearly prodiction to the proportion of this countries GDP within years.
