==> notebooks/3_GDP_adjusted_produciton.ipynb <==
import pandas as pd

gdp_data = pd.read_csv('../data/API_NY.GDP.MKTP.CD_DS2_en_csv_v2_740286/API_NY.GDP.MKTP.CD_DS2_en_csv_v2_740286.csv', header=2)
gdp_data.head()


# COULD HAVE AVOIDED THIS STEP IF STATED USING CODES OF COUNTRIES INSTEAD !
# 22 countries names to fix  'Bahamas, The',  'Brunei Darussalam', 'Cabo Verde', 
# 'Congo, Rep.',  'Congo, Dem. Rep.', 'Egypt, Arab Rep.',  'Gambia, The',
# 'Hong Kong SAR, China',  'Iran, Islamic Rep.',  'Macao SAR, China',
# 'Micronesia, Fed. Sts.',  'Russian Federation', 'St. Kitts and Nevis',
# 'St. Lucia', 'St. Martin (French part)', 'St. Vincent and the Grenadines',
# "Korea, Dem. People's Rep.",  'Korea, Rep.', 'Syrian Arab Republic',
# 'Turkiye',  'Venezuela, RB',  'Viet Nam',  'Yemen, Rep.',

gdp_data.loc[gdp_data['Country Name'] == 'Bahamas, The', 'Country Name'] = 'Bahamas'
gdp_data.loc[gdp_data['Country Name'] == 'Brunei Darussalam', 'Country Name'] = 'Brunei'
gdp_data.loc[gdp_data['Country Name'] == 'Cabo Verde', 'Country Name'] = 'Cape Verde'
gdp_data.loc[gdp_data['Country Name'] == 'Congo, Rep.', 'Country Name'] = 'Congo'
gdp_data.loc[gdp_data['Country Name'] == 'Congo, Dem. Rep.', 'Country Name'] = 'Democratic Republic of Congo'
gdp_data.loc[gdp_data['Country Name'] == 'Egypt, Arab Rep.', 'Country Name'] = 'Egypt'
gdp_data.loc[gdp_data['Country Name'] == 'Gambia, The', 'Country Name'] = 'Gambia'
gdp_data.loc[gdp_data['Country Name'] == 'Hong Kong SAR, China', 'Country Name'] = 'Hong Kong'
gdp_data.loc[gdp_data['Country Name'] == 'Iran, Islamic Rep.', 'Country Name'] = 'Iran'
gdp_data.loc[gdp_data['Country Name'] == 'Macao SAR, China', 'Country Name'] = 'Macao'
gdp_data.loc[gdp_data['Country Name'] == 'Micronesia, Fed. Sts.', 'Country Name'] = 'Micronesia (country)'
gdp_data.loc[gdp_data['Country Name'] == 'Korea, Dem. People\'s Rep.', 'Country Name'] = 'North Korea'
gdp_data.loc[gdp_data['Country Name'] == 'Russian Federation', 'Country Name'] = 'Russia'
gdp_data.loc[gdp_data['Country Name'] == 'St. Kitts and Nevis', 'Country Name'] = 'Saint Kitts and Nevis'
gdp_data.loc[gdp_data['Country Name'] == 'St. Lucia', 'Country Name'] = 'Saint Lucia'
gdp_data.loc[gdp_data['Country Name'] == 'St. Vincent and the Grenadines', 'Country Name'] = 'Saint Vincent and the Grenadines'
gdp_data.loc[gdp_data['Country Name'] == 'Korea, Rep.', 'Country Name'] = 'South Korea'
gdp_data.loc[gdp_data['Country Name'] == 'Syrian Arab Republic', 'Country Name'] = 'Syria'
gdp_data.loc[gdp_data['Country Name'] == 'Turkiye', 'Country Name'] = 'Turkey'
gdp_data.loc[gdp_data['Country Name'] == 'Venezuela, RB', 'Country Name'] = 'Venezuela'
gdp_data.loc[gdp_data['Country Name'] == 'Viet Nam', 'Country Name'] = 'Vietnam'
gdp_data.loc[gdp_data['Country Name'] == 'Yemen, Rep.', 'Country Name'] = 'Yemen'

set_gdp = set(gdp_data['Country Name'])

df_1 = pd.read_csv('../data/01_world_spatio_temporal_plastic_production_v0.csv')
df_1.columns = ['year', 'plastic_produced', 'country']
df_1.head()

set_plastic = set(df_1.country)

# understanding what countries have GDP data and which contries don't and why
common_countries = set_plastic.intersection(set_gdp)
absent_countries = set_plastic.difference(set_gdp)

print(f'Total countries in GDP dataset: {len(set_gdp)}')
print(f'Common countries: {len(common_countries)}')
print(f'Countries not in the list: {len(absent_countries)}')
absent_countries

### Adjusting total world production to the right GDP proportions

gdp_data_of_interest = gdp_data[gdp_data['Country Name'].isin(common_countries)]
gdp_data_of_interest.head()

countries_with_no_2010GDP = gdp_data_of_interest[gdp_data_of_interest['2010'].isna()]['Country Name']
gdp_data_of_interest[gdp_data_of_interest['2010'].isna()]

set(countries_with_no_2010GDP)

print(f'Before we throw countries with no GDP: {len(common_countries)}')
common_countries = common_countries.difference(countries_with_no_2010GDP)
print(f'After we throw countries with no GDP: {len(common_countries)}')

# re-creating countries of interest GDP
gdp_data_of_interest = gdp_data[gdp_data['Country Name'].isin(common_countries)]
gdp_data_of_interest.head()

# obtaining coefficient of countries of interest GDP to world GDP
x_to_world = gdp_data_of_interest['2010'].sum() / gdp_data['2010'].sum()
x_to_world

Value seems to small. Lets investigate more why that might be the case

print(f'Proportion of countries of interest to the whole dataset: {len(gdp_data_of_interest)/len(gdp_data)}')
print(len(set_gdp.difference(common_countries)))
print(set_gdp.difference(common_countries))

Okay, in this case we see that here are some aggregated entities such as European Union and North America, so double counting does exist in this case. We shouldn't worry about low proportion to the rest of the world... But we should worry about it, as we need to clean the dataset to understand if what proportion of the world plastic production is attributed to these countries... Let's see

set_gdp.difference(common_countries)

# we should remove these entities from the original dataset to avoid double counting
entiries_to_remove = { 'Africa Eastern and Southern',
 'Africa Western and Central', 'Arab World',
 'Early-demographic dividend',
 'East Asia & Pacific',
 'East Asia & Pacific (IDA & IBRD countries)',
 'East Asia & Pacific (excluding high income)',
 'Euro area',
 'Europe & Central Asia',
 'Europe & Central Asia (IDA & IBRD countries)',
 'Europe & Central Asia (excluding high income)',
 'European Union',
 'Fragile and conflict affected situations',
 'Heavily indebted poor countries (HIPC)',
 'High income',
 'IBRD only',
 'IDA & IBRD total',
 'IDA blend',
 'IDA only',
 'IDA total', 'Lao PDR',
 'Late-demographic dividend',
 'Latin America & Caribbean',
 'Latin America & Caribbean (excluding high income)',
 'Latin America & the Caribbean (IDA & IBRD countries)',
 'Least developed countries: UN classification', 'Low & middle income',
 'Low income',
 'Lower middle income', 'Lesotho',
 'Middle East & North Africa',
 'Middle East & North Africa (IDA & IBRD countries)',
 'Middle East & North Africa (excluding high income)',
 'Middle income', 'North America', 'Not classified',
 'OECD members',
 'Other small states',
 'Pacific island small states', 'Post-demographic dividend',
 'Pre-demographic dividend', 'Small states',
 'South Asia',
 'South Asia (IDA & IBRD)', 'Sub-Saharan Africa',
 'Sub-Saharan Africa (IDA & IBRD countries)',
 'Sub-Saharan Africa (excluding high income)', 'Upper middle income',
 'World',
}

print(f'Before we throw specified entities: {len(set_gdp)}')
set_gdp = set_gdp.difference(entiries_to_remove)
print(f'After we throw specified entities: {len(set_gdp)}')

# re - creating country gdp & re-calculating the percentage 
gdp_data = gdp_data[gdp_data['Country Name'].isin(set_gdp)]

x_to_world = gdp_data_of_interest['2010'].sum() / gdp_data['2010'].sum()
x_to_world

Now seems legit. Now we can created adjusted plastic production

from owid import catalog
# opening TheWorldBanks's catalogue
plastic_catalogs = catalog.find('plastic')
dataset_name = "plastic_use_polymer"
dataset_path = plastic_catalogs[plastic_catalogs['table'] == dataset_name].path.iloc[0]

# loading method from the docs
rc = catalog.RemoteCatalog() 
plastic_polymer_df = rc[dataset_path]

# adjusting total_plastic_prod_x
total_plastic_prod_x = plastic_polymer_df.loc['World', 2010]['total'] * x_to_world
total_plastic_prod_x

### Adjusting dataset with waste dataset with GDP

waste_generation = pd.read_csv('../data/plastic-waste-generation-total.csv')
waste_generation.columns = ['country', 'code', 'year', 'total_plastic_waste']
waste_generation

waste_generation.nunique()

168 * len(years_columns)

waste_to_production = waste_generation['total_plastic_waste'].sum() / total_plastic_prod_x

yearly_plastic_production_df = pd.DataFrame(columns=['country', 'code', 'year', 'total_plastic_prod'])

years_columns = list(set(gdp_data_of_interest.columns).difference({'Country Name', 'Country Code',
                                              'Indicator Name', 'Indicator Code',
                                              'Unnamed: 68'}))
years_columns.sort()
years_columns = years_columns[30:]

for country in common_countries:
    gdp_country = gdp_data_of_interest[gdp_data_of_interest['Country Name'] == country]
    base_value = gdp_country['2010']
    
    code = waste_generation.loc[waste_generation['country'] == country, 'code'].values[0]
    waste_2010 = waste_generation.loc[waste_generation['country'] == country, 'total_plastic_waste'].values[0]
    
    for year in years_columns:
        proprtion_to_2010 = (gdp_country[year]/base_value).values[0]
        plastic_production = waste_2010 * proprtion_to_2010 / waste_to_production
        data = [[country, code, year, plastic_production]]
        yearly_plastic_production_df = pd.concat([yearly_plastic_production_df, 
                                                pd.DataFrame(data,columns=['country', 'code', 'year', 'total_plastic_prod'])])

yearly_plastic_production_df

yearly_plastic_production_df[yearly_plastic_production_df.total_plastic_prod.isna()]

yearly_plastic_production_df.to_csv('../data/01_world_spatio_temporal_plastic_production_v1.csv')

Finished the first step adjusting plastic produciton to the csv 

### Trying to add regional boundaries
