==> notebooks/4_adjusted_interactive_map.ipynb <==
import pandas as pd
df_1 = pd.read_csv('../data/01_world_spatio_temporal_plastic_production_v1.csv')
df_1.columns = ['index',  'country', 'code', 'year', 'plastic_produced',]
df_1.head()

import geopandas as gpd
world = gpd.read_file('../data/ne_110m_admin_0_countries/ne_110m_admin_0_countries.shp')

#United States of America
world.loc[world['NAME'] == "Dem. Rep. Congo", 'NAME'] = 'Democratic Republic of Congo'
world.loc[world['NAME'] == "Eq. Guinea", 'NAME'] = 'Equatorial Guinea'
world.loc[world['NAME'] == "United States of America", 'NAME'] = 'United States'

set_01 = set(world.NAME)
set_02 = set(df_1.country)
len(set_01.intersection(set_02))

We lost some values (25% or 44 countries), but it's okay. Later we will figure out why there are some geometries missing.

# merging
gdf = world.merge(df_1, left_on='NAME', right_on='country')

# saving
gdf.to_csv('../data/02_world_spatio_temporal_plastic_production_v0.csv')

Some columns in world are redundunt. We need to remove them (maybe all except geometry,index, NAME)

import plotly.express as px
# Create an animated choropleth map
fig = px.choropleth(gdf,
                    geojson=gdf.geometry,
                    locations=gdf.index,
                    color='plastic_produced',
                    hover_name='country',
                    hover_data=['plastic_produced'],
                    animation_frame='year',
                    projection='natural earth')

fig.update_geos(fitbounds="locations", visible=False)
fig.update_layout(margin={"r":0,"t":0,"l":0,"b":0})

fig.show()

TODO:

• [] Work on the colorscheme of the visualized data

• [] Remove redundunt columns from the gdf
